DTO configurado

+++ src/main/java/com/contratosc/crud/services/ProductService.java
@@ -1,13 +1,17 @@
 package com.contratosc.crud.services;
 
+//#region imports
 import java.util.List;
 import java.util.Optional;
-
+import java.util.stream.Collectors;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.contratosc.crud.model.Product;
 import com.contratosc.crud.repository.ProductRepository;
+import com.contratosc.crud.shared.ProductDTO;
+//#endregion
 
 @Service
 public class ProductService {
@@ -15,25 +19,63 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
     
-    public List<Product> findAll(){
-        return productRepository.findAll();
+    public List<ProductDTO> findAll(){
+        List<Product>  products = productRepository.findAll();
+
+        return products.stream()
+        .map(product -> new ModelMapper().map(product, ProductDTO.class))
+        .collect(Collectors.toList());
     }
 
-    public Optional<Product> findById(Integer id){
-        return productRepository.findById(id);
+    public Optional<ProductDTO> findById(Integer id){
+
+        //Busca Produto pelo ID
+        Optional<Product> product = productRepository.findById(id);
+
+        //Exception Caso o produto não exista
+
+        // Converte o Product em ProductDTO
+        ProductDTO productDTO = new ModelMapper().map(product.get(), ProductDTO.class);
+
+        //Retorna Optional de ProductDTO
+        return Optional.of(productDTO);
     }
 
-    public Product addProduct(Product product){
-        return productRepository.save(product);
+    public ProductDTO addProduct(ProductDTO productDTO){
+
+        productDTO.setId(null);
+
+        Product product = new ModelMapper().map(productDTO, Product.class);
+
+        product = productRepository.save(product);
+
+        productDTO.setId(product.getId());
+
+        return productDTO;
     }
 
     public void deleteProduct(Integer id){
-        productRepository.deleteById(id);
+
+        Optional<Product> product = productRepository.findById(id);
+
+        if(product.isEmpty()){
+            System.out.println("Produto não encontrado");
+        }
+        else{
+            productRepository.deleteById(id);
+        }
+
+        
     }
 
-    public Product updateProduct(Integer id, Product product){
-        product.setId(id);
-        return productRepository.save(product);
+    public ProductDTO updateProduct(Integer id, ProductDTO productDTO){
+
+        productDTO.setId(id);
+
+        Product product = new ModelMapper().map(productDTO, Product.class);
+        productRepository.save(product);
+
+        return productDTO;
     }
 
 }
